Add libusb_open_device_with_fd: Create a handle for an already open file descriptor
Obtain a device handle for a device you've already opened yourself and thus
have a file-descriptor for.  This handle allows you to perform I/O on the
device in question.

The provided device handle will have no associated libusb_device, therefore
libusb_get_device() will return NULL on handles created with this function.

This is a non-blocking function; no requests are sent over the bus.

This function is only implemented on Linux, but it may be possible to add
implementations for other Unix like operating systems.

It will be useful for udev integration on Linux.  My particular use-case
is that I have a udev-rule which identifies a device creating a symlink
e.g. `/dev/example`.  I want a way to get a libusb handle to this
device.  Specifically I'm attempting to fix this bug in lirc using this
new API: https://sourceforge.net/p/lirc/tickets/186/ .

I believe that Android needs a way of doing this too as their
sandboxing technologies makes opening the device directly not possible.
Instead you can only get a hold of a USB device file-descriptor.

This is based on a patch from the chromium project written by Reilly Grant
<reillyg@chromium.org>.  For the original patch (based on v1.0.17) see:

* https://chromium.googlesource.com/chromium/src/third_party/+/c45fa9f8b58a170d15df6223a3a33381a948e59c
* https://codereview.chromium.org/1170853004

That patch has been forward-ported and reworked based on the concept
introduced in the previous commit of a libusb_device_handle which does
not have an associated libusb_device.  This allows clients to use the
I/O parts of libusb to be used independently of the enumeration parts.