Add libusb_open_fd: Create a handle for an already open file descriptor
Open a device and obtain a device handle. A handle allows you to perform
I/O on the device in question.

Instead of opening the device itself this function accepts an open file
descriptor that it will take ownership of.

Internally, this function adds a reference to the device and makes it
available to you through libusb_get_device(). This reference is removed
during libusb_close().

This is a non-blocking function; no requests are sent over the bus.

This function is only implemented on Linux, but it should be possible to
add implementations for other Unix like operating systems.

This patch originates from the chromium project and has been forward-
ported from v1.0.17 by William Manley.  The original author is Reilly
Grant <reillyg@chromium.org>.  For more information see:

* https://chromium.googlesource.com/chromium/src/third_party/+/c45fa9f8b58a170d15df6223a3a33381a948e59c
* https://codereview.chromium.org/1170853004

It will be useful for udev integration on Linux.  My particular use-case
is that I have a udev-rule which identifies a device creating a symlink
e.g. `/dev/example`.  I want a way to get a libusb handle to this
device.  I know that Android needs a way of doing this too as their
sandboxing technologies makes opening the device directly not possible.

[open-fd.patch]: https://chromium.googlesource.com/chromium/src/third_party/+/master/libusb/open-fd.patch